# express/static/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"Fruits": {
				"properties": {
					"name": {
						"type": "string"
					},
					"color": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"color"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"keycloak": {
				"type": "apiKey",
				"name": "Authorization",
				"description": "Keycloak Bearer Token",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/fruits": {
			"get": {
				"operationId": "GetFruit",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"properties": {
											"color": {
												"type": "string"
											},
											"name": {
												"type": "string"
											},
											"Id": {
												"type": "number",
												"format": "double"
											}
										},
										"required": [
											"color",
											"name",
											"Id"
										],
										"type": "object"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateFruit",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"color": {
											"type": "string"
										},
										"name": {
											"type": "string"
										},
										"Id": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"color",
										"name",
										"Id"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [
					{
						"keycloak": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Fruits"
							}
						}
					}
				}
			}
		},
		"/fruits/{id}": {
			"get": {
				"operationId": "GetFruitById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"color": {
													"type": "string"
												},
												"name": {
													"type": "string"
												},
												"Id": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"color",
												"name",
												"Id"
											],
											"type": "object"
										},
										{
											"type": "string",
											"enum": [
												"ไม่พบข้อมูล"
											]
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateFruit",
				"responses": {
					"200": {
						"description": "Updated",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"color": {
											"type": "string"
										},
										"name": {
											"type": "string"
										},
										"Id": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"color",
										"name",
										"Id"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [
					{
						"keycloak": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Fruits"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteFruit",
				"responses": {
					"200": {
						"description": "Deleted",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [
					{
						"keycloak": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/employees": {
			"get": {
				"operationId": "GetEmployees",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"properties": {
											"TitleOfCourtesy": {
												"type": "string"
											},
											"Title": {
												"type": "string"
											},
											"ReportsTo": {
												"type": "number",
												"format": "double"
											},
											"Region": {
												"type": "string"
											},
											"PostalCode": {
												"type": "string"
											},
											"PhotoPath": {
												"type": "string"
											},
											"Photo": {
												"type": "string"
											},
											"Notes": {
												"type": "string"
											},
											"LastName": {
												"type": "string"
											},
											"HomePhone": {
												"type": "string"
											},
											"HireDate": {
												"type": "string"
											},
											"FirstName": {
												"type": "string"
											},
											"Extension": {
												"type": "string"
											},
											"Country": {
												"type": "string"
											},
											"City": {
												"type": "string"
											},
											"BirthDate": {
												"type": "string"
											},
											"Address": {
												"type": "string"
											},
											"Id": {
												"type": "number",
												"format": "double"
											}
										},
										"required": [
											"TitleOfCourtesy",
											"Title",
											"ReportsTo",
											"Region",
											"PostalCode",
											"PhotoPath",
											"Photo",
											"Notes",
											"LastName",
											"HomePhone",
											"HireDate",
											"FirstName",
											"Extension",
											"Country",
											"City",
											"BirthDate",
											"Address",
											"Id"
										],
										"type": "object"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"keycloak": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateEmployee",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"TitleOfCourtesy": {
											"type": "string"
										},
										"Title": {
											"type": "string"
										},
										"ReportsTo": {
											"type": "number",
											"format": "double"
										},
										"Region": {
											"type": "string"
										},
										"PostalCode": {
											"type": "string"
										},
										"PhotoPath": {
											"type": "string"
										},
										"Photo": {
											"type": "string"
										},
										"Notes": {
											"type": "string"
										},
										"LastName": {
											"type": "string"
										},
										"HomePhone": {
											"type": "string"
										},
										"HireDate": {
											"type": "string"
										},
										"FirstName": {
											"type": "string"
										},
										"Extension": {
											"type": "string"
										},
										"Country": {
											"type": "string"
										},
										"City": {
											"type": "string"
										},
										"BirthDate": {
											"type": "string"
										},
										"Address": {
											"type": "string"
										},
										"Id": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"TitleOfCourtesy",
										"Title",
										"ReportsTo",
										"Region",
										"PostalCode",
										"PhotoPath",
										"Photo",
										"Notes",
										"LastName",
										"HomePhone",
										"HireDate",
										"FirstName",
										"Extension",
										"Country",
										"City",
										"BirthDate",
										"Address",
										"Id"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [
					{
						"keycloak": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"TitleOfCourtesy": {
										"type": "string"
									},
									"Title": {
										"type": "string"
									},
									"ReportsTo": {
										"type": "number",
										"format": "double"
									},
									"Region": {
										"type": "string"
									},
									"PostalCode": {
										"type": "string"
									},
									"PhotoPath": {
										"type": "string"
									},
									"Photo": {
										"type": "string"
									},
									"Notes": {
										"type": "string"
									},
									"LastName": {
										"type": "string"
									},
									"HomePhone": {
										"type": "string"
									},
									"HireDate": {
										"type": "string"
									},
									"FirstName": {
										"type": "string"
									},
									"Extension": {
										"type": "string"
									},
									"Country": {
										"type": "string"
									},
									"City": {
										"type": "string"
									},
									"BirthDate": {
										"type": "string"
									},
									"Address": {
										"type": "string"
									}
								},
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/employees/{id}": {
			"get": {
				"operationId": "GetEmployeeById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"TitleOfCourtesy": {
													"type": "string"
												},
												"Title": {
													"type": "string"
												},
												"ReportsTo": {
													"type": "number",
													"format": "double"
												},
												"Region": {
													"type": "string"
												},
												"PostalCode": {
													"type": "string"
												},
												"PhotoPath": {
													"type": "string"
												},
												"Photo": {
													"type": "string"
												},
												"Notes": {
													"type": "string"
												},
												"LastName": {
													"type": "string"
												},
												"HomePhone": {
													"type": "string"
												},
												"HireDate": {
													"type": "string"
												},
												"FirstName": {
													"type": "string"
												},
												"Extension": {
													"type": "string"
												},
												"Country": {
													"type": "string"
												},
												"City": {
													"type": "string"
												},
												"BirthDate": {
													"type": "string"
												},
												"Address": {
													"type": "string"
												},
												"Id": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"TitleOfCourtesy",
												"Title",
												"ReportsTo",
												"Region",
												"PostalCode",
												"PhotoPath",
												"Photo",
												"Notes",
												"LastName",
												"HomePhone",
												"HireDate",
												"FirstName",
												"Extension",
												"Country",
												"City",
												"BirthDate",
												"Address",
												"Id"
											],
											"type": "object"
										},
										{
											"type": "string",
											"enum": [
												"ไม่พบข้อมูล"
											]
										}
									]
								}
							}
						}
					}
				},
				"security": [
					{
						"keycloak": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateEmployee",
				"responses": {
					"200": {
						"description": "Updated",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"TitleOfCourtesy": {
											"type": "string"
										},
										"Title": {
											"type": "string"
										},
										"ReportsTo": {
											"type": "number",
											"format": "double"
										},
										"Region": {
											"type": "string"
										},
										"PostalCode": {
											"type": "string"
										},
										"PhotoPath": {
											"type": "string"
										},
										"Photo": {
											"type": "string"
										},
										"Notes": {
											"type": "string"
										},
										"LastName": {
											"type": "string"
										},
										"HomePhone": {
											"type": "string"
										},
										"HireDate": {
											"type": "string"
										},
										"FirstName": {
											"type": "string"
										},
										"Extension": {
											"type": "string"
										},
										"Country": {
											"type": "string"
										},
										"City": {
											"type": "string"
										},
										"BirthDate": {
											"type": "string"
										},
										"Address": {
											"type": "string"
										},
										"Id": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"TitleOfCourtesy",
										"Title",
										"ReportsTo",
										"Region",
										"PostalCode",
										"PhotoPath",
										"Photo",
										"Notes",
										"LastName",
										"HomePhone",
										"HireDate",
										"FirstName",
										"Extension",
										"Country",
										"City",
										"BirthDate",
										"Address",
										"Id"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [
					{
						"keycloak": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"TitleOfCourtesy": {
										"type": "string"
									},
									"Title": {
										"type": "string"
									},
									"ReportsTo": {
										"type": "number",
										"format": "double"
									},
									"Region": {
										"type": "string"
									},
									"PostalCode": {
										"type": "string"
									},
									"PhotoPath": {
										"type": "string"
									},
									"Photo": {
										"type": "string"
									},
									"Notes": {
										"type": "string"
									},
									"LastName": {
										"type": "string"
									},
									"HomePhone": {
										"type": "string"
									},
									"HireDate": {
										"type": "string"
									},
									"FirstName": {
										"type": "string"
									},
									"Extension": {
										"type": "string"
									},
									"Country": {
										"type": "string"
									},
									"City": {
										"type": "string"
									},
									"BirthDate": {
										"type": "string"
									},
									"Address": {
										"type": "string"
									}
								},
								"type": "object"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteEmployee",
				"responses": {
					"200": {
						"description": "Deleted",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [
					{
						"keycloak": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}